application/requirements.txt
serial
pandas
matplotlib.pyplot
struct
numpy
numpy.lib.function_base
datetime
os